slight file restructure
